Big refactoring, reintroducing state-arguments, moving lots of code from old to OpenXbox-comparable locations, all the while keeping it all compiling
