+++ src/devices/video/EmuNV2A_PTIMER.cpp
@@ -18,34 +18,34 @@ static inline uint64_t muldiv64(uint64_t a, uint32_t b, uint32_t c)
 }
 
 /* PIMTER - time measurement and time-based alarms */
-static uint32_t ptimer_get_clock()
+static uint32_t ptimer_get_clock(NV2AState * d)
 {
 	// Get time in nanoseconds
 	long int time = static_cast<long int>(std::chrono::duration_cast<std::chrono::nanoseconds>(std::chrono::high_resolution_clock::now().time_since_epoch()).count());
 	
-	return muldiv64(time, pramdac.core_clock_freq * ptimer.numerator, CLOCKS_PER_SEC * ptimer.denominator);
+	return muldiv64(time, d->pramdac.core_clock_freq * d->ptimer.numerator, CLOCKS_PER_SEC * d->ptimer.denominator);
 }
 
 DEVICE_READ32(PTIMER)
 {
 	DEVICE_READ32_SWITCH() {
 	case NV_PTIMER_INTR_0:
-		result = ptimer.pending_interrupts;
+		result = d->ptimer.pending_interrupts;
 		break;
 	case NV_PTIMER_INTR_EN_0:
-		result = ptimer.enabled_interrupts;
+		result = d->ptimer.enabled_interrupts;
 		break;
 	case NV_PTIMER_NUMERATOR:
-		result = ptimer.numerator;
+		result = d->ptimer.numerator;
 		break;
 	case NV_PTIMER_DENOMINATOR:
-		result = ptimer.denominator;
+		result = d->ptimer.denominator;
 		break;
 	case NV_PTIMER_TIME_0:
-		result = (ptimer_get_clock() & 0x7ffffff) << 5;
+		result = (ptimer_get_clock(d) & 0x7ffffff) << 5;
 		break;
 	case NV_PTIMER_TIME_1:
-		result = (ptimer_get_clock() >> 27) & 0x1fffffff;
+		result = (ptimer_get_clock(d) >> 27) & 0x1fffffff;
 		break;
 	default: 
 		result = 0;
@@ -62,21 +62,21 @@ DEVICE_WRITE32(PTIMER)
 	switch (addr) {
 
 	case NV_PTIMER_INTR_0:
-		ptimer.pending_interrupts &= ~value;
-		update_irq();
+		d->ptimer.pending_interrupts &= ~value;
+		update_irq(d);
 		break;
 	case NV_PTIMER_INTR_EN_0:
-		ptimer.enabled_interrupts = value;
-		update_irq();
+		d->ptimer.enabled_interrupts = value;
+		update_irq(d);
 		break;
 	case NV_PTIMER_DENOMINATOR:
-		ptimer.denominator = value;
+		d->ptimer.denominator = value;
 		break;
 	case NV_PTIMER_NUMERATOR:
-		ptimer.numerator = value;
+		d->ptimer.numerator = value;
 		break;
 	case NV_PTIMER_ALARM_0:
-		ptimer.alarm_time = value;
+		d->ptimer.alarm_time = value;
 		break;
 	default: 
 		//DEVICE_WRITE32_REG(ptimer); // Was : DEBUG_WRITE32_UNHANDLED(PTIMER);
