+++ src/devices/video/EmuNV2A_USER.cpp
@@ -3,9 +3,9 @@ DEVICE_READ32(USER)
 	unsigned int channel_id = addr >> 16;
 	assert(channel_id < NV2A_NUM_CHANNELS);
 
-	ChannelControl *control = &user.channel_control[channel_id];
+	ChannelControl *control = &d->user.channel_control[channel_id];
 
-	uint32_t channel_modes = pfifo.regs[NV_PFIFO_MODE];
+	uint32_t channel_modes = d->pfifo.regs[NV_PFIFO_MODE];
 
 	/* PIO Mode */
 	if (!channel_modes & (1 << channel_id)) {
@@ -37,17 +37,17 @@ DEVICE_WRITE32(USER)
 	unsigned int channel_id = addr >> 16;
 	assert(channel_id < NV2A_NUM_CHANNELS);
 
-	ChannelControl *control = &user.channel_control[channel_id];
+	ChannelControl *control = &d->user.channel_control[channel_id];
 
-	uint32_t channel_modes = pfifo.regs[NV_PFIFO_MODE];
+	uint32_t channel_modes = d->pfifo.regs[NV_PFIFO_MODE];
 	if (channel_modes & (1 << channel_id)) {
 		/* DMA Mode */
 		switch (addr & 0xFFFF) {
 		case NV_USER_DMA_PUT:
 			control->dma_put = value;
 
-			if (pfifo.cache1.push_enabled) {
-				pfifo_run_pusher();
+			if (d->pfifo.cache1.push_enabled) {
+				pfifo_run_pusher(d);
 			}
 			break;
 		case NV_USER_DMA_GET:
