+++ src/devices/video/nv2a.h
@@ -34,10 +34,503 @@
 // ******************************************************************
 #pragma once
 
+#include "Cxbx.h" // For xbaddr
 #include "devices\PCIDevice.h" // For PCIDevice
+
+#include <mutex>
+#include <condition_variable>
+#include <queue>
+#include <thread>
+
+#include <GL/glew.h>
+
+#include "nv2a_int.h"
 
 #define NV2A_ADDR  0xFD000000
 #define NV2A_SIZE             0x01000000
+
+#define NV_PMC_SIZE                 0x001000
+#define _NV_PFIFO_SIZE               0x002000 // Underscore prefix to prevent clash with NV_PFIFO_SIZE
+#define NV_PVIDEO_SIZE              0x001000
+#define NV_PTIMER_SIZE              0x001000
+#define NV_PFB_SIZE                 0x001000
+#define NV_PGRAPH_SIZE              0x002000
+#define NV_PCRTC_SIZE               0x001000
+#define NV_PRAMDAC_SIZE             0x001000
+#define NV_PRAMIN_ADDR   0x00700000
+#define NV_PRAMIN_SIZE              0x100000
+
+typedef xbaddr hwaddr; // Compatibility; Cxbx uses xbaddr, xqemu and OpenXbox use hwaddr 
+typedef uint32_t value_t; // Compatibility; Cxbx values are uint32_t (xqemu and OpenXbox use uint64_t)
+
+
+#define USE_TEXTURE_CACHE
+
+// Public Domain ffs Implementation
+// See: http://snipplr.com/view/22147/stringsh-implementation/
+constexpr int ffs(int v)
+{
+	unsigned int x = v;
+	int c = 1;
+
+	/*
+	* adapted from from
+	*      http://graphics.stanford.edu/~seander/bithacks.html#ZerosOnRightBinSearch
+	*
+	* a modified binary search algorithm to count 0 bits from
+	*  the right (lsb).  This algorithm should work regardless
+	*  of the size of ints on the platform.
+	*
+	*/
+
+	/* a couple special cases */
+	if (x == 0) return 0;
+	if (x & 1)  return 1;   /* probably pretty common */
+
+	c = 1;
+	while ((x & 0xffff) == 0) {
+		x >>= 16;
+		c += 16;
+	}
+	if ((x & 0xff) == 0) {
+		x >>= 8;
+		c += 8;
+	}
+	if ((x & 0x0f) == 0) {
+		x >>= 4;
+		c += 4;
+	}
+	if ((x & 0x03) == 0) {
+		x >>= 2;
+		c += 2;
+	}
+
+	c -= (x & 1);
+	c += 1;     /* ffs() requires indexing bits from 1 */
+				/*   ie., the lsb is bit 1, not bit 0  */
+	return c;
+}
+
+inline int GET_MASK(int v, int mask) {
+	return (((v) & (mask)) >> (ffs(mask) - 1));
+};
+
+inline int SET_MASK(int v, int mask, int val) {  
+    const unsigned int __val = (val);                            
+    const unsigned int __mask = (mask);                          
+
+    (v) &= ~(__mask);                                            
+    return (v) |= ((__val) << (ffs(__mask)-1)) & (__mask);              
+}
+
+#define CASE_4(v, step)                                              \
+    case (v):                                                        \
+	case (v)+(step) :												 \
+	case (v)+(step) * 2:                                             \
+	case (v)+(step) * 3
+
+
+#define NV2A_DEVICE(obj) \
+    OBJECT_CHECK(NV2AState, (obj), "nv2a")
+
+//void reg_log_read(int block, hwaddr addr, uint64_t val);
+//void reg_log_write(int block, hwaddr addr, uint64_t val);
+
+enum FifoMode {
+	FIFO_PIO = 0,
+	FIFO_DMA = 1,
+};
+
+enum FIFOEngine {
+	ENGINE_SOFTWARE = 0,
+	ENGINE_GRAPHICS = 1,
+	ENGINE_DVD = 2,
+};
+
+typedef struct DMAObject {
+	unsigned int dma_class;
+	unsigned int dma_target;
+	xbaddr address;
+	xbaddr limit;
+} DMAObject;
+
+typedef struct VertexAttribute {
+	bool dma_select;
+	xbaddr offset;
+
+	/* inline arrays are packed in order?
+	* Need to pass the offset to converted attributes */
+	unsigned int inline_array_offset;
+
+	float inline_value[4];
+
+	unsigned int format;
+	unsigned int size; /* size of the data type */
+	unsigned int count; /* number of components */
+	uint32_t stride;
+
+	bool needs_conversion;
+	uint8_t *converted_buffer;
+	unsigned int converted_elements;
+	unsigned int converted_size;
+	unsigned int converted_count;
+
+	float *inline_buffer;
+
+	GLint gl_count;
+	GLenum gl_type;
+	GLboolean gl_normalize;
+
+	GLuint gl_converted_buffer;
+	GLuint gl_inline_buffer;
+} VertexAttribute;
+
+typedef struct Surface {
+	bool draw_dirty;
+	bool buffer_dirty;
+	bool write_enabled_cache;
+	unsigned int pitch;
+
+	xbaddr offset;
+} Surface;
+
+typedef struct SurfaceShape {
+	unsigned int z_format;
+	unsigned int color_format;
+	unsigned int zeta_format;
+	unsigned int log_width, log_height;
+	unsigned int clip_x, clip_y;
+	unsigned int clip_width, clip_height;
+	unsigned int anti_aliasing;
+} SurfaceShape;
+
+typedef struct TextureShape {
+	bool cubemap;
+	unsigned int dimensionality;
+	unsigned int color_format;
+	unsigned int levels;
+	unsigned int width, height, depth;
+
+	unsigned int min_mipmap_level, max_mipmap_level;
+	unsigned int pitch;
+} TextureShape;
+
+typedef struct TextureKey {
+	TextureShape state;
+	uint64_t data_hash;
+	uint8_t* texture_data;
+	uint8_t* palette_data;
+} TextureKey;
+
+typedef struct TextureBinding {
+	GLenum gl_target;
+	GLuint gl_texture;
+	unsigned int refcnt;
+} TextureBinding;
+
+typedef struct KelvinState {
+	xbaddr dma_notifies;
+	xbaddr dma_state;
+	xbaddr dma_semaphore;
+	unsigned int semaphore_offset;
+} KelvinState;
+
+typedef struct ContextSurfaces2DState {
+	xbaddr dma_image_source;
+	xbaddr dma_image_dest;
+	unsigned int color_format;
+	unsigned int source_pitch, dest_pitch;
+	xbaddr source_offset, dest_offset;
+
+} ContextSurfaces2DState;
+
+typedef struct ImageBlitState {
+	xbaddr context_surfaces;
+	unsigned int operation;
+	unsigned int in_x, in_y;
+	unsigned int out_x, out_y;
+	unsigned int width, height;
+
+} ImageBlitState;
+
+typedef struct GraphicsObject {
+	uint8_t graphics_class;
+	union {
+		ContextSurfaces2DState context_surfaces_2d;
+
+		ImageBlitState image_blit;
+
+		KelvinState kelvin;
+	} data;
+} GraphicsObject;
+
+typedef struct GraphicsSubchannel {
+	xbaddr object_instance;
+	GraphicsObject object;
+	uint32_t object_cache[5];
+} GraphicsSubchannel;
+
+typedef struct GraphicsContext {
+	bool channel_3d;
+	unsigned int subchannel;
+} GraphicsContext;
+
+
+typedef struct PGRAPHState {
+	std::mutex mutex;
+
+	uint32_t pending_interrupts;
+	uint32_t enabled_interrupts;
+	std::condition_variable interrupt_cond;
+
+	xbaddr context_table;
+	xbaddr context_address;
+
+
+	unsigned int trapped_method;
+	unsigned int trapped_subchannel;
+	unsigned int trapped_channel_id;
+	uint32_t trapped_data[2];
+	uint32_t notify_source;
+
+	bool fifo_access;
+	std::condition_variable fifo_access_cond;
+	std::condition_variable flip_3d;
+
+	unsigned int channel_id;
+	bool channel_valid;
+	GraphicsContext context[NV2A_NUM_CHANNELS];
+
+	xbaddr dma_color, dma_zeta;
+	Surface surface_color, surface_zeta;
+	unsigned int surface_type;
+	SurfaceShape surface_shape;
+	SurfaceShape last_surface_shape;
+
+	xbaddr dma_a, dma_b;
+	//GLruCache *texture_cache;
+	bool texture_dirty[NV2A_MAX_TEXTURES];
+	TextureBinding *texture_binding[NV2A_MAX_TEXTURES];
+
+	//GHashTable *shader_cache;
+	//ShaderBinding *shader_binding;
+
+	bool texture_matrix_enable[NV2A_MAX_TEXTURES];
+
+	/* FIXME: Move to NV_PGRAPH_BUMPMAT... */
+	float bump_env_matrix[NV2A_MAX_TEXTURES - 1][4]; /* 3 allowed stages with 2x2 matrix each */
+
+	// wglContext *gl_context;
+	GLuint gl_framebuffer;
+	GLuint gl_color_buffer, gl_zeta_buffer;
+	GraphicsSubchannel subchannel_data[NV2A_NUM_SUBCHANNELS];
+
+	xbaddr dma_report;
+	xbaddr report_offset;
+	bool zpass_pixel_count_enable;
+	unsigned int zpass_pixel_count_result;
+	unsigned int gl_zpass_pixel_count_query_count;
+	GLuint* gl_zpass_pixel_count_queries;
+
+	xbaddr dma_vertex_a, dma_vertex_b;
+
+	unsigned int primitive_mode;
+
+	bool enable_vertex_program_write;
+
+	//uint32_t program_data[NV2A_MAX_TRANSFORM_PROGRAM_LENGTH][VSH_TOKEN_SIZE];
+
+	uint32_t vsh_constants[NV2A_VERTEXSHADER_CONSTANTS][4];
+	bool vsh_constants_dirty[NV2A_VERTEXSHADER_CONSTANTS];
+
+	/* lighting constant arrays */
+	uint32_t ltctxa[NV2A_LTCTXA_COUNT][4];
+	bool ltctxa_dirty[NV2A_LTCTXA_COUNT];
+	uint32_t ltctxb[NV2A_LTCTXB_COUNT][4];
+	bool ltctxb_dirty[NV2A_LTCTXB_COUNT];
+	uint32_t ltc1[NV2A_LTC1_COUNT][4];
+	bool ltc1_dirty[NV2A_LTC1_COUNT];
+
+	// should figure out where these are in lighting context
+	float light_infinite_half_vector[NV2A_MAX_LIGHTS][3];
+	float light_infinite_direction[NV2A_MAX_LIGHTS][3];
+	float light_local_position[NV2A_MAX_LIGHTS][3];
+	float light_local_attenuation[NV2A_MAX_LIGHTS][3];
+
+	//VertexAttribute vertex_attributes[NV2A_VERTEXSHADER_ATTRIBUTES];
+
+	unsigned int inline_array_length;
+	uint32_t inline_array[NV2A_MAX_BATCH_LENGTH];
+	GLuint gl_inline_array_buffer;
+
+	unsigned int inline_elements_length;
+	uint32_t inline_elements[NV2A_MAX_BATCH_LENGTH];
+
+	unsigned int inline_buffer_length;
+
+	unsigned int draw_arrays_length;
+	unsigned int draw_arrays_max_count;
+
+	/* FIXME: Unknown size, possibly endless, 1000 will do for now */
+	GLint gl_draw_arrays_start[1000];
+	GLsizei gl_draw_arrays_count[1000];
+
+	GLuint gl_element_buffer;
+	GLuint gl_memory_buffer;
+	GLuint gl_vertex_array;
+
+	uint32_t regs[NV_PGRAPH_SIZE]; // TODO : union
+} PGRAPHState;
+
+
+typedef struct CacheEntry {
+	//QSIMPLEQ_ENTRY(CacheEntry) entry;
+	unsigned int method : 14;
+	unsigned int subchannel : 3;
+	bool nonincreasing;
+	uint32_t parameter;
+} CacheEntry;
+
+typedef struct Cache1State {
+	unsigned int channel_id;
+	FifoMode mode;
+
+	/* Pusher state */
+	bool push_enabled;
+	bool dma_push_enabled;
+	bool dma_push_suspended;
+	xbaddr dma_instance;
+
+	bool method_nonincreasing;
+	unsigned int method : 14;
+	unsigned int subchannel : 3;
+	unsigned int method_count : 24;
+	uint32_t dcount;
+
+	bool subroutine_active;
+	xbaddr subroutine_return;
+	xbaddr get_jmp_shadow;
+	uint32_t rsvd_shadow;
+	uint32_t data_shadow;
+	uint32_t error;
+
+	bool pull_enabled;
+	enum FIFOEngine bound_engines[NV2A_NUM_SUBCHANNELS];
+	enum FIFOEngine last_engine;
+
+	/* The actual command queue */
+	std::mutex mutex;
+	std::condition_variable cache_cond;
+	std::queue<CacheEntry*> cache;
+	std::queue<CacheEntry*> working_cache;
+} Cache1State;
+
+typedef struct ChannelControl {
+	xbaddr dma_put;
+	xbaddr dma_get;
+	uint32_t ref;
+} ChannelControl;
+
+typedef struct NV2AState {
+    // PCIDevice dev;
+    // qemu_irq irq;
+    bool exiting;
+
+    // VGACommonState vga;
+    // GraphicHwOps hw_ops;
+    // QEMUTimer *vblank_timer;
+
+    // MemoryRegion *vram;
+    // MemoryRegion vram_pci;
+    uint8_t *vram_ptr;
+    size_t vram_size;
+    // MemoryRegion ramin;
+	struct {
+		uint8_t *ramin_ptr;
+		size_t ramin_size;
+		uint32_t regs[NV_PRAMIN_SIZE]; // TODO : union
+	} pramin;
+
+    // MemoryRegion mmio;
+    // MemoryRegion block_mmio[NV_NUM_BLOCKS];
+
+    struct {
+        uint32_t pending_interrupts;
+        uint32_t enabled_interrupts;
+		uint32_t regs[NV_PMC_SIZE]; // TODO : union
+    } pmc;
+
+    struct {
+		std::thread puller_thread;
+        uint32_t pending_interrupts;
+        uint32_t enabled_interrupts;
+        Cache1State cache1;
+		uint32_t regs[_NV_PFIFO_SIZE]; // TODO : union
+    } pfifo;
+
+    struct {
+		uint32_t regs[NV_PVIDEO_SIZE]; // TODO : union
+    } pvideo;
+
+    struct {
+        uint32_t pending_interrupts;
+        uint32_t enabled_interrupts;
+        uint32_t numerator;
+        uint32_t denominator;
+        uint32_t alarm_time;
+		uint32_t regs[NV_PTIMER_SIZE]; // TODO : union
+    } ptimer;
+
+    struct {
+		uint32_t regs[NV_PFB_SIZE]; // TODO : union
+    } pfb;
+
+    struct PGRAPHState pgraph;
+
+    struct {
+        uint32_t pending_interrupts;
+        uint32_t enabled_interrupts;
+        hwaddr start;
+		uint32_t regs[NV_PCRTC_SIZE]; // TODO : union
+    } pcrtc;
+
+    struct {
+        uint32_t core_clock_coeff;
+        uint64_t core_clock_freq;
+        uint32_t memory_clock_coeff;
+        uint32_t video_clock_coeff;
+		uint32_t regs[NV_PRAMDAC_SIZE]; // TODO : union
+    } pramdac;
+
+    struct {
+        ChannelControl channel_control[NV2A_NUM_CHANNELS];
+    } user;
+
+	// PRMCIO (Actually the VGA controller)
+	struct {
+		uint8_t cr_index;
+		uint8_t cr[256]; /* CRT registers */
+	} prmcio;
+
+    std::mutex io_lock;
+    //SDL_Window *sdl_window;
+
+} NV2AState;
+
+typedef value_t(*read_func)(NV2AState *d, hwaddr addr); //, unsigned int size);
+typedef void(*write_func)(NV2AState *d, hwaddr addr, value_t val); //, unsigned int size);
+
+typedef struct {
+	read_func read;
+	write_func write;
+} MemoryRegionOps;
+
+typedef struct NV2ABlockInfo {
+	const char* name;
+	hwaddr offset;
+	uint64_t size;
+	MemoryRegionOps ops;
+} NV2ABlockInfo;
 
 class NV2ADevice : public PCIDevice {
 public:
